Refactor convert command and final changes
